% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDistance.R
\name{analyzeDistance}
\alias{analyzeDistance}
\title{Calculate the distance along and away from a user-defined boundary using mFISH cell coordinates}
\usage{
analyzeDistance(
  SectionName,
  coord_data,
  highlight_cluster.1 = NULL,
  highlight_cluster.2 = NULL,
  highlight_colour = "black"
)
}
\arguments{
\item{SectionName}{Name of your section (example: 'section7'), used for plotting titles and saving filenames}

\item{coord_data}{Coordinate section data extracted from mFISH object using getCoords().}

\item{highlight_cluster.1}{Specify identity of first cluster to highlight when defining boundary, to use as a landmark if needed (example: highlight_cluster.1 = 1).}

\item{highlight_cluster.2}{Specify identity of second cluster to highlight when defining boundary, to use as a landmark if needed (example: highlight_cluster.1 = 2).}

\item{highlight_colour}{Specify colour to highlight cells belonging to selected cluster in 'highlight_cluster.1'. Default colour is black}
}
\value{
A dataframe with cell ID, coordinates, cluster ID, distance along and distance away from your defined boundary. Distance metrics are available as raw (microns) and normalized (0-1) outputs.
}
\description{
Compute distance along and the distance away from a user-defined boundary, using section-specific coordinate data computed with getCoords function. Useful to analyze cell distances away from white matteer, or neocortical layer edge.
}
\author{
Margarita Kapustina
}
