% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAtlas.R
\name{clusterAtlas}
\alias{clusterAtlas}
\title{mFISH Atlas Clustering}
\usage{
clusterAtlas(bins, k, density = 0, include.density = F, normalize = F)
}
\arguments{
\item{bins}{A bin dataframe generated from: \code{createAtlas}}

\item{k}{Number of clusters to create (via hierarchical clustering of binned windows)}

\item{density}{Density value to use as a cut-off. If left at 0, automatically chooses optimal cut-off value.}

\item{include.density}{If you wish to include cell body density as a value to cluster on, select TRUE.}

\item{normalize}{Set to true if you wish to normalize cell body density.}
}
\value{
A bin dataframe with a cluster column. Can be plotted using \code{plotAtlas}.
}
\description{
A function that creates data-driven anatomical atlas parcellation by taking a binned dataframe from createAtlas() and clusters based on cellular phenotype and/or cellular density.
}
\author{
Kaitlin Sullivan
}
