% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneBoxPlot.R
\name{geneBoxPlot}
\alias{geneBoxPlot}
\title{Create a box plot for a given gene}
\usage{
geneBoxPlot(mFISH, gene, group = "cluster")
}
\arguments{
\item{mFISH}{An mFISH object}

\item{gene}{Gene name to plot}

\item{group}{Metadata variable to group data by (must be a factor)}
}
\value{
A box plot showing gene expression
}
\description{
Create a box plot for a given gene
}
\author{
Kaitlin E Sullivan
}
