% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFISH.R
\name{goFISH}
\alias{goFISH}
\title{Launches the Gone mFISHing shiny app}
\usage{
goFISH(mFISH, filter.by = NA, k = NA)
}
\arguments{
\item{filter.by}{A vector of strings or single string value of a gene to filter the data by}

\item{k}{A numeric value denoting number of clusters to input}

\item{mfish}{An mFISH object}
}
\value{
Launch Gone mFISHing shiny app
}
\description{
Launches the Gone mFISHing shiny app
}
\author{
Kaitlin E Sullivan
}
