% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load mFISH data (written by Mark Cembrowski)}
\usage{
loadData(anm = c(), reg = c(), excThres = 1e-05, nClust = 4, perplexity = 30)
}
\arguments{
\item{anm}{\emph{character vector.} If length>0, select only the animals present in this vector.}

\item{reg}{\emph{character vector.} If length>0, select only the regions ant','int','pos' present in this vector.}

\item{excThres}{\emph{numeric} Slc17a7 cutoff value for removing non-excitatory cells.}

\item{nClust}{\emph{numeric} Number of clusters}

\item{perplexity}{\emph{numeric} Perplexity of tSNE visualization.}
}
\value{
A normalized dataframe containing the quantified mFISH.
}
\description{
Load mFISH data (written by Mark Cembrowski)
}
