% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAdd.R
\name{metaAdd}
\alias{metaAdd}
\title{mFISH Add Metadata}
\usage{
metaAdd(mfish_object, metadata_to_add, metadata_variables_to_add)
}
\arguments{
\item{mfish_object}{mFISH object}

\item{metadata_to_add}{New metadata as a dataframe (must contain cell IDs in an 'id' column).}

\item{metadata_variables_to_add}{Specify variable (column name) to add to mFISH object as new metadata.}
}
\value{
mFISH object with updated metadata.
}
\description{
Add metadata to an existing mFISH object. (Note: new metadata must contain cell IDs.)
}
\author{
Margarita Kapustina
}
