% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFISH.R
\name{newFISH}
\alias{newFISH}
\title{Launches the Gone mFISHing shiny app}
\usage{
newFISH(mFISH, filter.by = NA)
}
\arguments{
\item{mfish}{An mFISH object}
}
\value{
Launch Gone mFISHing shiny app
}
\description{
Launches the Gone mFISHing shiny app
}
\author{
Kaitlin E Sullivan
}
