% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCluster.R
\name{plotCluster}
\alias{plotCluster}
\title{Plot cluster identity (written by Mark Cembrowski)}
\usage{
plotCluster(xy = "geog")
}
\arguments{
\item{xy}{\emph{character} if =='umap', plot via UMAP dim reduction / if =='tsne', plot via TSNE dim reduction / if =='geog' (or anything else) plot via spatial (geographical) coords}
}
\value{
ggplot of clustered data
}
\description{
Plot cluster identity (written by Mark Cembrowski)
}
