% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDendro.R
\name{plotDendro}
\alias{plotDendro}
\title{mFISH Dendrogram Plot}
\usage{
plotDendro(mFISH, k = "current", polar = F)
}
\arguments{
\item{mFISH}{An mFISH object that has been run through \code{ruCluster}.}

\item{k}{At default "current", it will colour the dendrogram by the selected number of clusters. Note: if you change the number of clusters, the original cluster colours will change!}

\item{polar}{A logical indicating whether the plot will have polar coordinates.}
}
\value{
Plots a dendrogram coloured by the chosen number of clusters.
}
\description{
Plot and colour the dendrogram of one's hierarchical clustering. To be used after running \code{ruCluster} to guide one's choice of cluster number.
}
\author{
Kaitlin E Sullivan
}
