% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDim.R
\name{plotDim}
\alias{plotDim}
\title{Plot UMAP of mFISH data}
\usage{
plotDim(mFISH, colour.by = "cluster", size = 0.9, cols = NA)
}
\arguments{
\item{mFISH}{An mFISH object with UMAP coordinates}

\item{colour.by}{A metadata variable to colour by}

\item{size}{Point size}

\item{cols}{Vector of colours}
}
\value{
A ggplot showing UMAP space coloured by metadata.
}
\description{
Plot UMAP of mFISH data
}
