% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene.R
\name{plotGene}
\alias{plotGene}
\title{Plot cluster identity (written by Mark Cembrowski)}
\usage{
plotGene(theGene, xy = "geog", doNorm = F, lo = -1, hi = -1)
}
\arguments{
\item{theGene}{\emph{character} gene to plot.}

\item{xy}{\emph{character} if =='umap', plot via UMAP dim reduction / if =='tsne', plot via TSNE dim reduction / if =='geog' (or anything else) plot via spatial (geographical) coords}

\item{doNorm}{\emph{logical} Plot normalized value?}

\item{lo}{\emph{numeric} If >0, sets low value of plotted range.}

\item{hi}{\emph{numeric} If >0, sets high value of plotted range.}
}
\value{
ggplot of clustered data
}
\description{
Plot cluster identity (written by Mark Cembrowski)
}
