% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVar.R
\name{plotVar}
\alias{plotVar}
\title{mFISH PCA Variance Plot}
\usage{
plotVar(mFISH, metric = "variance", include.auto = T)
}
\arguments{
\item{mFISH}{An mFISH object that has undergone processing via \code{ruProcess()}.}

\item{metric}{A PCA metric to plot. Options are: "variance", "cumulative", or "stdev". This corresponds to the proportion of variance, cumulativie proportion, or standard deviation of a given PC.}

\item{include.auto}{A logical indicator to include a vertical assymptote at the number of PCs auto-selected. Note: you can change number of PCs used in dim reduction and clustering in \code{ruUMAP}.}
}
\value{
An elbow plot for the mFISH principle components.
}
\description{
Plot the proportion of variance or standard deviation of each principle component to determine the number of PCs to be used in subsequent analysis. Use this function after \code{ruProcess} and before \code{ruUMAP}.
}
\author{
Kaitlin Sullivan
}
