% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\name{plotViolin}
\alias{plotViolin}
\title{Plot violin of gene expression (written by Mark Cembrowski)}
\usage{
plotViolin(theGene, doNorm = F, alpha = 0.2)
}
\arguments{
\item{theGene}{\emph{character} Gene to plot}

\item{doNorm}{\emph{logical} Plot normalized value?}

\item{alpha}{\emph{numeric} ransparency of individual cell data points. If <=0, turns individual points off. If =1, points full opaque.}
}
\value{
ggplot of clustered data
}
\description{
Plot violin of gene expression (written by Mark Cembrowski)
}
