% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawPixels.R
\name{rawPixels}
\alias{rawPixels}
\title{Create a dataframe with raw pixel info}
\usage{
rawPixels()
}
\value{
A dataframe containing the raw pixels mFISH data of all channels present.
}
\description{
Create a dataframe with raw pixel info
}
