% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruFilter.R
\name{ruFilter}
\alias{ruFilter}
\title{Select features and filter your mFISH object by a given gene}
\usage{
ruFilter(mFISH, threshold = 0.1, filter.by = NA, exclude = NA)
}
\arguments{
\item{mFISH}{An mFISH object}

\item{threshold}{A numeric value to threshold the gene by}

\item{filter.by}{A gene or a vector of gene names to filter data by}

\item{exclude}{A gene or vector of gene names to exclude from analysis}
}
\value{
An mFISH object with a populated filteredData section
}
\description{
Select features and filter your mFISH object by a given gene
}
\author{
Kaitlin E Sullivan
}
