% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruMake.R
\name{ruMake}
\alias{ruMake}
\title{Create a dataframe containing quantified mFISH data}
\usage{
ruMake(dir, region = NA, id = NA)
}
\arguments{
\item{dir}{Directory containing the quantified channel tables.}

\item{region}{A character string with the name of the tissue region imaged.}

\item{id}{Animal number or unique identifier for tissue sample.}
}
\value{
A dataframe containing the quantified mFISH data of all channels present.
}
\description{
Create a dataframe containing quantified mFISH data
}
