% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruNormal.R
\name{ruNormal}
\alias{ruNormal}
\title{Normalize and filter an mFISH dataset}
\usage{
ruNormal(table, filter = NA, value = 0)
}
\arguments{
\item{table}{Table made by the ruMake() function.}

\item{filter}{A gene name to filter data by before normalizing.}

\item{value}{A count-per-area value to be used in the filtering step}
}
\value{
A normalized dataframe containing the quantified mFISH.
}
\description{
Normalize and filter an mFISH dataset
}
