% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruProcess.R
\name{ruProcess}
\alias{ruProcess}
\title{Normalizes data and runs a PCA}
\usage{
ruProcess(mFISH, remove.outliers = F, outlier.thresh = c(1, 11))
}
\arguments{
\item{mFISH}{An mFISH object containing raw or filtered data}

\item{remove.outliers}{A logical value for removing "outliers" (eg: autofluorescent cells positive for all genes)}

\item{outlier.thresh}{Low and high threshold for outliers. These numbers will change based on amount of genes and autofluorescence in your data.}
}
\value{
and mFISH object, normalized and containing a PCA (found in \code{obj@attributes$pca})
}
\description{
Normalizes data and runs a PCA
}
\author{
Kaitlin E Sullivan
}
