% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruUMAP.R
\name{ruUMAP}
\alias{ruUMAP}
\title{mFISH UMAP Dimensionality Reduction}
\usage{
ruUMAP(mFISH, metric = "manhattan", nn = 15, min.dist = 0.1, npc = "auto")
}
\arguments{
\item{mFISH}{An mFISH object with preprocessing completed.}

\item{metric}{Distance metric for UMAP.}

\item{nn}{Nearest neighbours value for UMAP.}

\item{min.dist}{Minimum distance value for UMAP.}

\item{npc}{Number of principle components to use.}
}
\value{
An mFISH object with UMAP coordinates populated in the metadata slot.
}
\description{
Run UMAP dimensionality reduction on your mFISH object. To be used after \code{ruProcess} and before \code{ruCluster}.
}
\author{
Kaitlin E Sullivan
}
