% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterBoxPlot.R
\name{clusterBoxPlot}
\alias{clusterBoxPlot}
\title{mFISH Cluster Box Plot}
\usage{
clusterBoxPlot(mFISH, clus = "all", metadata = "genes")
}
\arguments{
\item{mFISH}{An mFISH object}

\item{clus}{Cluster name or number in quotations (or "all" for every cluster)}

\item{metadata}{Metadata variable to plot (must be numeric)}
}
\value{
A box plot showing gene expression per cluster
}
\description{
Create a box plot denoting expression of \strong{all genes} in \emph{one or all clusters}.
}
\author{
Kaitlin E Sullivan
}
