% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneBoxPlot.R
\name{geneBoxPlot}
\alias{geneBoxPlot}
\title{mFISH Gene Box Plot}
\usage{
geneBoxPlot(mFISH, gene, group = "cluster")
}
\arguments{
\item{mFISH}{An mFISH object}

\item{gene}{Gene name to plot}

\item{group}{Metadata variable to group data by (must be a factor)}
}
\value{
A box plot showing gene expression
}
\description{
Create a boxplot denoting expression of a \strong{single gene} across \emph{all clusters}.
}
\author{
Kaitlin E Sullivan
}
