% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFISH.R
\name{goFISH}
\alias{goFISH}
\title{Launches the Gone mFISHing shiny app}
\usage{
goFISH(table)
}
\arguments{
\item{table}{A table with values created via ruMake()}
}
\value{
Launch Gone mFISHing shiny app
}
\description{
Launches the Gone mFISHing shiny app
}
