% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFISH.R
\name{goFISH}
\alias{goFISH}
\title{Gone mFISHing}
\usage{
goFISH(mFISH, filter.by = NA, k = NA, norm = "PAC")
}
\arguments{
\item{mFISH}{An mFISH object}

\item{filter.by}{A vector of strings or single string value of a gene to filter the data by}

\item{k}{A numeric value denoting number of clusters to input}

\item{norm}{Normalization method of "PAC" or "log"}
}
\value{
Launch Gone mFISHing shiny app
}
\description{
Launch a Shiny app for code-free analysis of mFISH data.
}
\author{
Kaitlin E Sullivan
}
