% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDim.R
\name{plotDim}
\alias{plotDim}
\title{mFISH UMAP Plot}
\usage{
plotDim(mFISH, colour.by = "cluster", size = 0.9, cols = NA)
}
\arguments{
\item{mFISH}{An mFISH object that has undergone all analysis steps}

\item{colour.by}{A metadata variable to colour by (auto-coloured by cluster).}

\item{size}{Point size}

\item{cols}{Vector of colours}
}
\value{
A ggplot showing UMAP space coloured by metadata.
}
\description{
Plot UMAP dimensionality reduction and colour by gene expression or metadata.
}
\author{
Kaitlin E Sullivan
}
