% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpace.R
\name{plotSpace}
\alias{plotSpace}
\title{mFISH Plotting Cells in Native Tissue}
\usage{
plotSpace(
  mFISH,
  colour.by = "cluster",
  size = 0.9,
  cols = NA,
  include.fil = T,
  group.by = NA
)
}
\arguments{
\item{mFISH}{An mFISH object that has undergone all analysis steps.}

\item{colour.by}{A metadata variable to colour by}

\item{size}{Point size}

\item{cols}{Vector of colours}

\item{include.fil}{Logical allowing for filtered cells to appear in grey}

\item{group.by}{A variable to separate plots by (eg: section or anum to view individual slices or animals)}
}
\value{
A ggplot showing cells in native space!
}
\description{
Plot cells in their original X,Y coordinates. Colour according to gene expression or metadata.
}
\author{
Kaitlin E Sullivan
}
