% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruCluster.R
\name{ruCluster}
\alias{ruCluster}
\title{mFISH Clustering}
\usage{
ruCluster(mFISH, k, dmetric = "euclidean", p = 2)
}
\arguments{
\item{mFISH}{An mFISH object}

\item{k}{Number of clusters}

\item{dmetric}{Metric for creating distance matrix (see \code{dist()})}

\item{p}{Power of the Minkowski distance (see \code{dist()})}
}
\value{
A clustered mFISH object.
}
\description{
Hierarchically cluster an mFISH object. To be used after \code{ruUMAP}.
}
\author{
Kaitlin E Sullivan
}
