% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruCombine.R
\name{ruCombine}
\alias{ruCombine}
\title{mFISH Data Pooling}
\usage{
ruCombine(dfs)
}
\arguments{
\item{dfs}{A list of dataframes (from \code{ruRead} using \code{list(df1,df2,df3)})}
}
\value{
A combined dataframe.
}
\description{
Combine \code{ruRead} datatables from multiple mFISH images!
}
\author{
Kaitlin E Sullivan
}
