% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruRead.R
\name{ruRead}
\alias{ruRead}
\title{Read in mFISH data from FIJI}
\usage{
ruRead(dir, region = NA, anum = NA, section = NA)
}
\arguments{
\item{dir}{Directory containing the quantified channel tables.}

\item{region}{A character string with the name of the tissue region imaged.}

\item{anum}{Animal number or unique identifier for tissue sample.}

\item{section}{Section number}
}
\value{
A dataframe containing the quantified mFISH data of all channels present.
}
\description{
Create a combined datatable from individual gene quantification tables generated by FIJI. Use \code{ruCombine} to combine multiple \code{ruRead} tables before \emph{creating and mFISH object} with \code{ruMake}.
}
\author{
Kaitlin E Sullivan
}
